#287 split finite volume tests into unit and convergence
